New push

+++ src/main/java/com/yoramu/controller/PatientCont.java
@@ -9,6 +9,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -27,9 +28,6 @@ import java.util.stream.IntStream;
 public class PatientCont {
     @Autowired
     PatientService patientService;
-    public static String UPLOAD_DIRECTORY = System.getProperty("user.dir") + "/src/main/resources/static/images";
-//    @Value("${upload.directory}")
-//    private String UPLOAD_DIRECTORY;
 
 //    @GetMapping
 //    String patientGetterSlash(Model model, HttpSession session){
@@ -47,13 +45,16 @@ public class PatientCont {
     @GetMapping
     public String listPatients(Model model, @RequestParam("page") Optional<Integer> page, @RequestParam("size") Optional<Integer> size, @RequestParam(value = "search", required = false) String search) {
         currentPage(model, page, size, search);
-
+        return "patients";
+    }
+    @GetMapping("/")
+    public String listPatientSlash(Model model, @RequestParam("page") Optional<Integer> page, @RequestParam("size") Optional<Integer> size, @RequestParam(value = "search", required = false) String search) {
+        currentPage(model, page, size, search);
         return "patients";
     }
     @GetMapping("/patients")
     public String listBooks(Model model, @RequestParam("page") Optional<Integer> page, @RequestParam("size") Optional<Integer> size, @RequestParam(value = "search", required = false) String search) {
         currentPage(model, page, size, search);
-
         return "patients";
     }
 
@@ -75,30 +76,18 @@ public class PatientCont {
     }
 
     @PostMapping("/addPatient")
-    String addPatient(@ModelAttribute Patient patient, @RequestParam("image") MultipartFile file, @RequestParam("action") String action, @RequestParam("pic") String img) throws IOException {
-        StringBuilder fileNames = new StringBuilder();
+    String addPatient(@ModelAttribute Patient patient, @RequestParam("action") String action, BindingResult br) throws IOException {
+        if(br.hasErrors()){
+            return "patients";
+        }
         if(action.equals("Update")){
             patient.setId(patient.getId());
-            if(file.getSize() != 0){
-                Path fileNameAndPath = Paths.get(UPLOAD_DIRECTORY, file.getOriginalFilename());
-                fileNames.append(file.getOriginalFilename());
-                Files.write(fileNameAndPath, file.getBytes());
-
-                patient.setPicture(fileNames.toString());
-            }else {
-                patient.setPicture(img);
-            }
             patientService.addPatient(patient);
         }else if(action.equals("Delete")){
             patient.setId(patient.getId());
             patientService.deletePatient(patient);
         }else{
-            Path fileNameAndPath = Paths.get(UPLOAD_DIRECTORY, file.getOriginalFilename());
-            fileNames.append(file.getOriginalFilename());
-            Files.write(fileNameAndPath, file.getBytes());
-
             patient.setId(UUID.randomUUID());
-            patient.setPicture(fileNames.toString());
             patientService.addPatient(patient);
         }
         return "redirect:patients";
